Initial commit of working nubsudge
This is pretty simple, it takes a single srt file as input and an
integer. It linearly interpolates the value so that the last subtitle of
the file has both timestamps moved by the value of the nudge. Every
other subtitle is moved based on what percentage it is of the total time
covered in the file. This is useful for correcting subtitle slew